# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nkap
  # Not checked against the POMs' own version, on purpose, not by oversight (issue #106
  # considered it and decided against, for this slice): OpenApiSpecIT already verifies
  # paths, methods, status codes and fields against the running application, which is what
  # a caller actually depends on -- this string has no consumer and no functional effect.
  # Wiring it to ${project.version} would mean introducing Maven resource filtering or the
  # spring-boot-maven-plugin's build-info goal into this build, neither configured today,
  # for a field a human already has open and edits by hand at the exact moment they change
  # every POM in the same release commit (CONTRIBUTING.md's own release sequence). Worth
  # automating the day this project releases often enough that a human is no longer reading
  # this file closely each time; not yet.
  #
  # Moved to "1.3.0" here, independent of whatever the POMs currently carry: this file
  # describes the API as published, and it changed again -- a second callback route,
  # POST /callbacks/{providerId}/{reference}, was added beside the existing one (issue #185)
  # so a callback can be attributed by the address the gateway chose rather than only by its
  # body. A new path is additive, so a minor bump, the same rule the previous "1.2.0" move
  # gives for itself: bump it only when a path, method, field or status code in this file
...
  # appVersion).
  version: "1.3.0"
  description: |
...

...

...

...
  license:
...
security:
  - bearerAuth: []
...
paths:
  /payments:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment and submit it to the operator
      description: |
        Creates a `COLLECT` (take money from `counterpartyMsisdn`) or `DISBURSE` (send money
        to it) and submits it before answering — there is no separate "create" and "submit"
        step, because a payment that exists but was never sent is a state this API does not
        have a use for.

        **The response status is the whole point of this endpoint.** `201 Created` means the
        request reached a definite answer without needing you to wait for one: most often
        because the operator answered — `SUBMITTED`, `PENDING`, or already a terminal state —
        but `state: FAILED` with `history[].cause: GATEWAY` on the last transition is `201`
        too, and means the opposite of "the operator answered": this gateway refused the
        request itself, from what it already knew, and never asked the operator anything (ADR
        0013) — an intent for an operation the addressed installation does not support, most
        commonly. Check `cause`, not the status code, to tell the two apart; `history` is
        always present so you always can. `202 Accepted` means neither of those: the operator
        was asked and did not answer in time, `state` is `UNKNOWN`, and that is not an error to
        retry, log as a failure, or tell a payer about. It means exactly what it says — the
        outcome is not yet known — and it resolves itself: Nkap's own reconciler keeps asking
        the operator on a backoff, and a webhook (`docs/webhooks.md`) fires the moment it
        reaches `SUCCEEDED`, `FAILED` or `EXPIRED`. Poll `GET /payments/{reference}` or wait
        for that webhook; do not resubmit the same intent under a new `Idempotency-Key`, which
        would be a second real payment.

        **`Idempotency-Key` is two-sided, not a dedup cache.** The same key with the same
        request body always replays the first response verbatim (`201`/`202`, unchanged) — no
        second submission reaches the operator, which is what makes retrying a request whose
        response was lost in transit safe. The same key with a *different* body is `409`: this
        API refuses to guess which of two different requests you meant, rather than silently
        picking one.

        **The merchant is the API key's, never a field in this body.** There is no
        `merchantId` here on purpose — see `CreatePaymentRequest`'s own description.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: >
            A definite answer without waiting: `state` is `SUBMITTED` or `PENDING` (the
            operator answered), already terminal for an operator that settles synchronously
            (the operator answered), or `FAILED` with the last `history[].cause` `GATEWAY` —
            this gateway refused the request itself and never asked the operator at all (ADR
            0013). See this operation's own description for how to tell the two apart.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentResponse" }
        "202":
          description: >
            The operator did not answer in time. `state` is `UNKNOWN` — not known, not
            failed. See this operation's own description for what to do next.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentResponse" }
        "400":
          description: >
            The request is malformed or names something this deployment cannot serve —
            `missing-idempotency-key`, `malformed-request` (unreadable body, or a fractional
            `amount`), `invalid-request` (a required field is missing or not a value the
            domain accepts), `unserved-currency` (this installation settles a different
            currency), or `unconfigured-country` (no installation answers this `country`).
            Nothing is created and no `Idempotency-Key` claim is left behind — the same key
            works again once the request is fixed.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409":
          description: >
            `idempotency-key-reuse` (the same key, a different body) or `request-in-progress`
            (an earlier request with this key has not finished yet — retry once it has).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /payments/{reference}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Read a payment's stored state
      description: |
        Reads what Nkap has on file. **This never calls the operator** — see the `payments`
        tag's own description for why — so it is instant and safe to poll as often as you
        like while a payment is `UNKNOWN`, `SUBMITTED` or `PENDING`.

        A payment belonging to another merchant answers exactly like one that does not exist
        at all (`404`, identical body): two different answers would be an oracle for which
        references exist under someone else's key.
      parameters:
        - $ref: "#/components/parameters/ReferencePath"
      responses:
        "200":
          description: The payment, as currently stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentResponse" }
        "400":
          description: "`malformed-reference` — the path segment is not a well-formed reference."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404":
          description: "`payment-not-found` — no such reference, or it belongs to another merchant."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "nkap_<43 base64url characters>"
      description: |
        `Authorization: Bearer <key>`. Keys are minted by a host-side command, never a route,
        and shown exactly once — `java -jar app.jar --nkap.apikey.create
        --nkap.apikey.merchant=<id>` (see `docs/integration-guide.md`, step 1). The key
        identifies the merchant; there is no separate `merchantId` anywhere in a request
        body. One boolean on the key, `admin`, gates operator-wide reads
        (`GET /balance`, `GET /statements/imports/{importId}`, both webhook admin routes) —
        not a role system.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        A key of the caller's choosing, unique per intended payment. The same key replays the
        first response; the same key with a different body is `409`. See this operation's own
        description for the full two-sided rule.
      schema: { type: string }
    ReferencePath:
      name: reference
      in: path
      required: true
      description: A payment's reference, returned by `POST /payments` or `POST /payments/{reference}/refunds`.
      schema: { type: string, format: uuid }
  headers:
    Location:
      description: "`/payments/{reference}` — the same value as the body's own `reference` field, as a URL."
      schema: { type: string }
  responses:
    Unauthenticated:
      description: >
        No API key was presented, or the one presented is not recognised. One sentence for
        both cases — a different message for "wrong key" than "no key" would confirm that
        valid keys exist, which is an oracle.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      description: >
        `application/problem+json` (RFC 7807). Every error this API returns uses this shape,
        from one place (`ApiExceptionHandler`) — a client sees the same envelope whatever went
        wrong. `type` is a stable identifier a client can branch on; it does not resolve to a
        document. See each operation's own `4xx`/`5xx` responses for which `type` means what.
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type: { type: string, format: uri, examples: ["https://nkap.dev/problems/payment-not-found"] }
        title: { type: string, description: "A short, stable summary of this problem type." }
        status: { type: integer, description: "Repeats the HTTP status code." }
        detail: { type: string, description: "What went wrong, specific to this request — safe to show a developer, not written for an end payer." }
        instance: { type: string, description: "The request path that produced this problem." }
    CreatePaymentRequest:
      description: >
        The body of `POST /payments`. **There is no `merchantId`.** The merchant is the one
        the API key identifies, never a value the caller supplies — see `bearerAuth`'s own
        description for why that used to be a real hole.
      type: object
      required: [operation, amount, currency, country, counterpartyMsisdn]
      properties:
        operation:
          type: string
          enum: [COLLECT, DISBURSE]
          description: Take money from `counterpartyMsisdn` (`COLLECT`) or send money to it (`DISBURSE`).
        amount:
          type: integer
          format: int64
          minimum: 1
          description: >
            An integer count of the currency's minor units — `5000` XAF is 5000 francs (XAF
            has none), `5000` EUR is €50.00. A fractional number (`"50.5"`) is a `400` before
            anything is persisted; this is never a decimal, anywhere in this API.
        currency:
          type: string
          description: >
            An ISO 4217 code this gateway counts in (`dev.nkap.core.money.Currency`) — not
            enumerated here because membership only grows; see `docs/providers/mtn.md` for
            what a given installation actually settles in.
        country:
          type: string
          description: >
            Which configured installation this payment routes to, e.g. `cm` for the `mtn-cm`
            installation — explicit rather than guessed from the MSISDN, the currency, or the
            merchant, each rejected for a way it could quietly charge the wrong installation.
            An unconfigured country is a `400` naming what is configured.
        counterpartyMsisdn:
          type: string
          description: The payer's (or payee's) number, in the format the operator expects.
        payerMessage:
          type: string
          description: Shown to the payer on their handset, where the operator supports it. Omitted, `""`.
        payeeNote:
          type: string
          description: A note attached to the transfer for the payee's own records. Omitted, `""`.
...
    PaymentResponse:
      description: >
        What `POST /payments`, `POST /payments/{reference}/refunds` and
        `GET /payments/{reference}` all return — a refund is a payment, read the same way.
        Every field is always present; a value that does not apply is `""`, never absent or
        `null` (`refundOf` on a payment that is not a refund, `providerTransactionId` before
        one exists).
      type: object
      required:
        - reference
        - provider
        - merchantId
        - state
        - operation
        - amountMinorUnits
        - currency
        - counterpartyMsisdn
        - providerReference
        - providerTransactionId
        - createdAt
        - updatedAt
        - detail
        - refundOf
        - escalatedAt
        - unresolvedSince
        - history
      properties:
        reference: { type: string, format: uuid }
        provider:
          type: string
          description: "The installation this payment routed to, e.g. `mtn-cm` — `mtn-` plus the request's own `country`."
        merchantId: { type: string, description: The merchant the API key that created this payment identifies. }
        state:
          type: string
          enum: [CREATED, SUBMITTED, PENDING, UNKNOWN, SUCCEEDED, FAILED, EXPIRED]
          description: >
            See `docs/integration-guide.md`'s step 3 for what to do with each one. `UNKNOWN`
            is never terminal; `SUCCEEDED`/`FAILED`/`EXPIRED` are, and each fires exactly one
            webhook (`docs/webhooks.md`).
        operation: { type: string, enum: [COLLECT, DISBURSE] }
        amountMinorUnits: { type: integer, format: int64 }
        currency: { type: string }
        counterpartyMsisdn: { type: string }
        providerReference: { type: string, description: "The operator's own reference for the request, once known. `\"\"` until then." }
        providerTransactionId:
          type: string
          description: >
            The operator's transaction id for the settled movement. `""` unless `state` is
            `SUCCEEDED` — and, against the bundled simulator specifically, `""` even then; see
            `docs/providers/mtn.md`.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        detail:
          type: string
          description: '`""` unless `state` is `UNKNOWN`, where it explains that the outcome is not yet known and names the poll URL.'
        refundOf: { type: string, description: "The original collection's reference, for a refund. `\"\"` for every payment that is not one." }
        escalatedAt:
          type: string
          description: >
            An RFC3339 timestamp for when the reconciler gave up chasing this payment
            automatically and handed it to a human, because the operator answered
            inconclusively past the escalation window — or `""` if this payment has never
            been escalated. Not a `state` value — see the `state` enum above — because
            escalation describes what this gateway did about the operator's silence, not
            what the operator said; an escalated payment still shows `PENDING` or `UNKNOWN`
            and still resolves by a later callback or query.
        unresolvedSince:
          type: string
          description: >
            An RFC3339 timestamp for when this payment last became unresolved — left
            `CREATED` for `SUBMITTED`, `PENDING` or `UNKNOWN` — unchanged by a later hop
            between those three, or `""` if this payment is not currently in that posture.
            Answers how long a caller has been waiting, not how many times this gateway has
            queried the operator about it (see `docs/integration-guide.md`'s step 3).
        history:
          type: array
          items: { $ref: "#/components/schemas/PaymentTransition" }
    PaymentTransition:
      type: object
      required: [from, to, at, cause, operatorCode, note]
      properties:
        from: { type: string, enum: [CREATED, SUBMITTED, PENDING, UNKNOWN, SUCCEEDED, FAILED, EXPIRED] }
        to: { type: string, enum: [CREATED, SUBMITTED, PENDING, UNKNOWN, SUCCEEDED, FAILED, EXPIRED] }
        at: { type: string, format: date-time }
        cause:
          type: string
          enum: [SUBMIT_RESPONSE, QUERY, CALLBACK, RECONCILER, GATEWAY]
          description: >
            What attributed this transition — the initial submit response, a live query, the
            operator's own callback, the reconciler's backoff loop, or `GATEWAY`: this
            gateway itself, deciding without asking the operator anything, because the
            request asked for an operation the adapter does not support or an adapter
            refused a currency mismatch before making any call. Membership only grows.
        operatorCode: { type: string, description: "The operator's own code for this transition, when it sent one. Often \"\"." }
        note: { type: string }
...